+++ backend/user-service/src/main/java/com/unikly/userservice/UserServiceApplication.java
@@ -2,9 +2,13 @@ package com.unikly.userservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.persistence.autoconfigure.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication(scanBasePackages = {"com.unikly.userservice", "com.unikly.common"})
+@EntityScan(basePackages = {"com.unikly.userservice", "com.unikly.common"})
+@EnableJpaRepositories(basePackages = {"com.unikly.userservice", "com.unikly.common"})
 @EnableScheduling
 public class UserServiceApplication {